pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skill-graph"
version = "0.1.0"
description = "Knowledge graph + vector search for AI agent skill discovery"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "numpy>=1.26",
    "falkordb>=1.0.0",
    "sentence-transformers>=2.2",
    "fastmcp>=0.1",
]

[project.scripts]
skill-graph = "skill_graph.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.hatch.build.targets.wheel]
packages = ["src/skill_graph"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
